Split Queue.js business logic and view logic
