+++ scripts/modules/Queue/Queue.js
@@ -1,0 +1,5 @@
+import QueueView from './queueView.js';
+
+export function init() {
+  QueueView();
+}
